change sp1 bindings to be u64 aligned

+++ operator/sp1/lib/sp1.h
@@ -1,4 +1,5 @@
 #include <stdbool.h>
+#include <stdint.h>
 
-bool verify_sp1_proof_ffi(unsigned char *proof_buffer, unsigned int proof_len,
-                          unsigned char *elf_buffer, unsigned int elf_len);
+bool verify_sp1_proof_ffi(unsigned char *proof_buffer, uint64_t proof_len,
+                          unsigned char *elf_buffer, uint64_t elf_len);
